+++ src/tools/fsTools/isAhk.ts
@@ -1,7 +1,8 @@
 import type * as vscode from 'vscode';
 
+const isAhkReg: Readonly<RegExp> = /\.ahk$/iu;
 export function isAhk(fsPath: string): boolean {
-    return (/\.ahk$/iu).test(fsPath);
+    return isAhkReg.test(fsPath);
 }
 
 export function isAhkTab(uri: vscode.Uri): boolean {